+++ app/flatpak-builtins.h
@@ -28,6 +28,7 @@
 #include "flatpak-tty-utils-private.h"
 #include "flatpak-utils-private.h"
 #include "flatpak-dir-private.h"
+#include "flatpak-xml-utils-private.h"
 
 G_BEGIN_DECLS
