+++ common/meson.build
@@ -197,6 +197,7 @@ sources = [
   'flatpak-utils-http.c',
   'flatpak-utils.c',
   'flatpak-uri.c',
+  'flatpak-xml-utils.c',
   'flatpak-zstd-decompressor.c',
 ]
